+++ MordoEngine/src/Scene/Game/GameScene.cpp
@@ -11,6 +11,23 @@ GameScene::GameScene(std::shared_ptr<Camera> camera, std::shared_ptr<TerrainSyst
 	m_SkyBoxRenderer = std::make_unique<SkyBoxRenderer>(
 		Manager<Shader>::Get("skyBox")
 	);
+
+	glm::vec3 centerTerrainPosition = m_TerrainSystem->GetMiddleTerrainPosition();
+	float yOffset = 0.2f;
+	float scale = m_TerrainSystem->GetTerrainWorldScale();
+
+	auto AddLight = [&](float offsetX, float offsetZ)
+		{
+			float x = centerTerrainPosition.x + offsetX * scale;
+			float z = centerTerrainPosition.z + offsetZ * scale;
+			float y = m_TerrainSystem->GetTerrainInterpolatedHeightAt(x, z, yOffset);
+			m_LightSystem->AddPointLight(PointLight(glm::vec3(x, y, z)));
+		};
+
+	AddLight(0.0f, 0.0f);
+	AddLight(50.0f, 50.0f);
+	AddLight(-100.0f, -100.0f);
+	AddLight(-100.0f, -100.0f);
 }
 
 void GameScene::OnEntry()
@@ -22,20 +39,20 @@ void GameScene::Update(float deltaTime)
 {
 	float velocity = 100.0f * m_TerrainSystem->GetTerrainWorldScale() * deltaTime;
 	m_CameraController->Update(deltaTime, velocity, *m_TerrainSystem);
-    m_LightSystem->Update(deltaTime);
+	m_LightSystem->Update(deltaTime);
 }
 
 void GameScene::Render()
 {
-    glm::mat4 projection = m_Camera->GetProjectionMatrix();
-    glm::mat4 view = m_Camera->GetViewMatrix();
-    glm::mat4 model = glm::mat4(1.0f);
-    glm::vec3 cameraPos = m_Camera->GetPosition();
+	glm::mat4 projection = m_Camera->GetProjectionMatrix();
+	glm::mat4 view = m_Camera->GetViewMatrix();
+	glm::mat4 model = glm::mat4(1.0f);
+	glm::vec3 cameraPos = m_Camera->GetPosition();
 
-    Shader& terrainShader = Manager<Shader>::Get("terrain");
-    m_LightSystem->Render(terrainShader, cameraPos, &projection, &view, &model);
-    m_TerrainSystem->Render(terrainShader, cameraPos, &projection, &view, &model);
-    m_SkyBoxRenderer->Render(&view, &projection, nullptr);
+	Shader& terrainShader = Manager<Shader>::Get("terrain");
+	m_LightSystem->Render(terrainShader, cameraPos, &projection, &view, &model);
+	m_TerrainSystem->Render(terrainShader, cameraPos, &projection, &view, &model);
+	m_SkyBoxRenderer->Render(&view, &projection, nullptr);
 }
 
 GameScene::~GameScene()